Added javadoc for LWJGL library
Changed the glOrtho viewport to a gluPerspective / gluLookAt for 3D rendering. Had to swap the Y-axis from the original design, but this makes everything much easier to read.

Made a rotation effect when pushing left mouse button it rotates left and when pressing right mouse button it rotates right. This is made to invert the controls in x direction.