Done some more cleaning in the code.
Changed the pacman.png and ghost image to have white silhuettes instead of specified colors. Now it's possible to define the colors in the game.


Needs to be fixed:
 - Make ghosts dark-blue when they not are immortal
 - Make ghost body disappear when it's killed (reappear when ghost's in the nest)
 - Make ghosts flash when they are warning





CAUTION

Some test are still *ucked-up, due to the refactoring and simplifications made (tests aren't working because of the structural changes)... Gonna rewrite the tests during the weekend, to make them more supporting and cover more of the code.