Done some more cleaning in the code. Find a better way of randomizing colors (now they are too similar... perhaps on a 0.5f randomization on all three colors would do).
Make sure that the ghosts don't move when the screen is moved... I could get all ghosts of screen now!



CAUTION

Some test are still *ucked-up, due to the refactoring and simplifications made (tests aren't working because of the structural changes)... Gonna rewrite the tests during the weekend, to make them more supporting and cover more of the code.