Changed colors in GhostColor to host a list of 10 default colors, to be used based on the ghostColorIndex. May add more clear-colors later if the need arises.
Added a method to GameLWJGL that can be overriden to prevent the game to update and render based on some conditions. That is used in PacmanGame, to stop the updating and rendering of the game, when the mouse is outside the drawarea or if the window looses it's focus. This stops the problem of having the game to update ghost positions when clicking the window bar (this made the ghosts able to leave the window).

Fixed respawn algorithm to force all ghosts to be inside the nest, when killed (this problem showed only, when there were more than 4 ghosts).


CAUTION

Some test are still *ucked-up, due to the refactoring and simplifications made (tests aren't working because of the structural changes)... Gonna rewrite the tests during the weekend, to make them more supporting and cover more of the code.